move patch_courses_test.go to testing package
